fitur upload profile namun belum berjalan

+++ backend/auth_service/models/user.go
@@ -2,7 +2,7 @@ package models
 
 type User struct {
 	// gorm.Model
-	ID        int64  `gorm:"primaryKey"`
+	ID        int32  `gorm:"primaryKey"`
 	Username  string `gorm:"unique;not null"`
 	Email     string `gorm:"unique;not null" json:"email" binding:"required,email"`
 	Password  string `gorm:"not null"`
